pyproject.toml
[tool.poetry]
name = "wk-delivery"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
celery = "^5.3.6"
python-dotenv = "^1.0.0"
sqlmodel = "^0.0.12"
redis = "^5.0.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
